Integrate formula expansion.
 - Moved into pyecsca.ec.formula subpackage.
 - Unified formula metrics.
 - Moved tests into tests with library_formula fixture.
 - Cleaned up typing.
